am62px: ipc: Fix Misra C Rule 10.3 issues
MISRA.ETYPE.ASSIGN.2012
Assignment to an object of a narrower essential type
or a different essential type category.

Fixes: SITSW-5606

+++ source/drivers/ipc_notify/v0/ipc_notify_v0.c
@@ -122,7 +122,7 @@ void IpcNotify_isr(void *args)
             for(msg=0; msg<numMsgs; msg++)
             {
                 value = IpcNotify_mailboxRead(mailboxBaseAddr, hwFifoId);
-                clientId = (value >> IPC_NOTIFY_CLIENT_ID_SHIFT) & (IPC_NOTIFY_CLIENT_ID_MAX - 1U);
+                clientId = (uint16_t)((value >> IPC_NOTIFY_CLIENT_ID_SHIFT) & (IPC_NOTIFY_CLIENT_ID_MAX - 1U));
 
                 if(gIpcNotifyCtrl.callback[clientId]!=NULL)
                 {
